seems to work
a hybrid of two different tutorials: http://elm-lang.org/examples/zip-codes and https://gist.github.com/freakingawesome/7f86ed7683cfeeec4557 we are able to write a responsive app that looks up the names of the repos for any user.

additional follow-up work could be to enter the names of repos and list the files.  also improve the layout if possible.  all of this and more to come.